% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploscores.R
\name{get.quickeffects}
\alias{get.quickeffects}
\title{Quick Allele Effects}
\usage{
get.quickeffects(chromo, start, end, allele.effects)
}
\arguments{
\item{chromo}{Integer, chromosome number.}

\item{start}{Numeric, locus start position (Mb).}

\item{end}{Numeric, locus end position (Mb).}

\item{allele.effects}{Data frame, describes allele effects (e.g. A.score = 0, B.score = 1)}
}
\value{
score.avg
}
\description{
Quickly enerate scores based on haplotype effects with no founder attributions
For strains with residual heterozygosity and/or recombinations
Starts by averaging all allele effects on same chromosome
Then averages chromosomes
}
