% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploscores.R
\name{haploscores2}
\alias{haploscores2}
\title{Haplotype Scores 2}
\usage{
haploscores2(chromo, start, end, allele.effects)
}
\arguments{
\item{chromo}{Integer, chromosome number.}

\item{start}{Numeric, locus start position (Mb).}

\item{end}{Numeric, locus end position (Mb).}

\item{allele.effects}{Data frame, describes allele effects (e.g. A.score = 0, B.score = 1)}
}
\value{
haplos
}
\description{
Generate scores based on haplotype effects only for animals
with no recombinations within the locus or heterozygosity
}
