% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploscores.R
\name{haploscores3}
\alias{haploscores3}
\title{Complex Haplotype Scores}
\usage{
haploscores3(chromo, start, end, allele.effects)
}
\arguments{
\item{chromo}{Integer, chromosome number.}

\item{start}{Numeric, locus start position (Mb).}

\item{end}{Numeric, locus end position (Mb).}

\item{allele.effects}{Data frame, describes allele effects (e.g. A.score = 0, B.score = 1)}
}
\value{
multi
}
\description{
Generate scores based on haplotype effects only for animals
with residual heterozygosity or recombinations within the locus
}
