% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploscores_2.R
\name{haplotypes}
\alias{haplotypes}
\title{Locus Haplotypes without allele effects}
\usage{
haplotypes(chromo, start, end)
}
\arguments{
\item{chromo}{Integer, chromosome number.}

\item{start}{Numeric, locus start position (Mb).}

\item{end}{Numeric, locus end position (Mb).}
}
\value{
Data frame containing strain, alias, chromosome,
haplotype, start_position, end_position, and founder
}
\description{
Assign CC strain haplotypes without locus scores
}
\examples{
locushaplos(11, 70.5, 72)

}
