% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line2alias.R
\name{line2alias}
\alias{line2alias}
\title{Line 2 Alias}
\usage{
line2alias(df, strain.col, is.rix = F)
}
\arguments{
\item{df}{Input data frame which contains a column with new CC line designations}

\item{strain.col}{Character, name of column containing strains to be changed- not required if is.rix=T.}

\item{is.rix}{Logical, is the data from inbred crosses (RIX)?}
}
\value{
Input data frame with CC line designation converted to old alias nomenclature
}
\description{
Convert from new CC0XX line designations to old CC nomenclature
}
