% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploscores.R
\name{locushaplos}
\alias{locushaplos}
\title{Locus Haplotype Scores}
\usage{
locushaplos(chromo, start, end)
}
\arguments{
\item{chromo}{Integer, chromosome number.}

\item{start}{Numeric, locus start position (Mb).}

\item{end}{Numeric, locus end position (Mb).}
}
\value{
Data frame containing strain, alias, chromosome,
haplotype, start_position, end_position, and founder
}
\description{
Get founder haplotype at a locus of interest
}
\examples{
locushaplos(1, 70.5, 72)

}
