% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate2table.R
\name{plate2table}
\alias{plate2table}
\title{Plate to table}
\usage{
plate2table(df, date = F)
}
\arguments{
\item{df}{Data frame}

\item{date}{Logical, is there a row at the bottom of the absorbance values with the date the assay was run}
}
\value{
df
}
\description{
Convert from horizontal platereader layout to vertical data frame
}
\examples{

}
