% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploscores.R
\name{quickhaplos}
\alias{quickhaplos}
\title{Quick Allele Effects}
\usage{
quickhaplos(chromo, start, end, allele.effects)
}
\arguments{
\item{chromo}{Integer, chromosome number.}

\item{start}{Numeric, locus start position (Mb).}

\item{end}{Numeric, locus end position (Mb).}

\item{allele.effects}{Data frame, describes allele effects (e.g. A.score = 0, B.score = 1...)}
}
\value{
Data frame containing strain, alias, and allele effects
}
\description{
Quickly generate scores based on haplotype effects with no founder attributions.
For strains with residual heterozygosity and/or recombinations, allele effects on the same
same chromosome are averaged, followed by averaging of the two chromosomes.
}
\examples{
allele.effects=c(
A.score <- 0 ,
B.score <- 1 ,
C.score <- 0 ,
D.score <- 1 ,
E.score <- 0 ,
F.score <- 0 ,
G.score <- 0 ,
H.score <- 1 )

quickeffects(chromo = 1, start = 70.5, end = 72)
}
