% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceref.R
\name{replaceast}
\alias{replaceast}
\title{Remove asterices}
\usage{
replaceast(df)
}
\arguments{
\item{df}{data frame of Sanger output}
}
\value{
Data frame with reference allele calls
}
\description{
Remove * from allele calls
}
\examples{
replaceast(merge_output)

}
