% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceref.R
\name{replaceref}
\alias{replaceref}
\title{Fill in reference calls}
\usage{
replaceref(df)
}
\arguments{
\item{df}{data frame of Sanger output}
}
\value{
Data frame with reference allele calls
}
\description{
Replaced reference calls (- or ~, keyed in as NA) with actual reference allele
}
\examples{
replaceref(merge_output)

}
