% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addhaplotype.R
\name{addhap}
\alias{addhap}
\title{Add haplotypes}
\usage{
addhap(df, haplos, locus.name, strain.col)
}
\arguments{
\item{haplos}{Data frame, contains haplotype assignments for each CC strain as generated by haploscores function (haploscores, haploscores2, haploscores2, or quickscores).}

\item{locus.name}{Character, name assigned to the locus (e.g. "chr1").}

\item{strain.col}{Character, name of column containing strains to be changed - not required if is.rix=T.}
}
\value{
df
}
\description{
Add haplotypes at a locus of interest to CC data frame
}
\examples{

}
