% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addhaplotype.R
\name{addmx1}
\alias{addmx1}
\title{Add Mx1}
\usage{
addmx1(df, haplos, locus.name, cast.score = 0.5)
}
\arguments{
\item{haplos}{Data frame, contains haplotype assignments for each CC strain as generated by haploscores.}

\item{locus.name}{Character, name assigned to the locus (e.g. "chr1").}

\item{cast.score}{Numeric (0, 0.5, or 1).}
}
\value{
df
}
\description{
Add Mx1 allele status to a CC data frame
}
\examples{

}
