% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploscores.R
\name{haploscores}
\alias{haploscores}
\title{Haplotype Scores}
\usage{
haploscores(chromo, start, end, allele.effects)
}
\arguments{
\item{chromo}{Integer, chromosome number.}

\item{start}{Numeric, locus start position (Mb).}

\item{end}{Numeric, locus end position (Mb).}

\item{allele.effects}{Data frame, describes allele effects (e.g. A.score = 0, B.score = 1)}
}
\value{
haplos
}
\description{
Generate scores based on haplotype effects
}
\examples{

}
