% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploscores.R
\name{haploscores2}
\alias{haploscores2}
\title{Haplotype Scores 2}
\usage{
haploscores2(chromo, start, end, allele.effects)
}
\arguments{
\item{chromo}{Integer, chromosome number.}

\item{start}{Numeric, locus start position (Mb).}

\item{end}{Numeric, locus end position (Mb).}

\item{allele.effects}{Data frame, describes allele effects (e.g. A.score = 0, B.score = 1...)}
}
\value{
Data frame containing strain, alias, chromosome, haplotype, start_position, end_position, allele effect score, and founder
}
\description{
Associate allele effect scores with each CC strain based on founder haplotype,
only for CC strains with NO recombinations within the locus and
NO residual heterozygosity
}
\examples{
allele.effects=c(
A.score <- 0 ,
B.score <- 1 ,
C.score <- 0 ,
D.score <- 1 ,
E.score <- 0 ,
F.score <- 0 ,
G.score <- 0 ,
H.score <- 1 )

haploscores2(chromo = 1, start = 70.5, end = 72)
}
