% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploscores.R
\name{locushaplos}
\alias{locushaplos}
\title{Locus Haplotype Scores}
\usage{
locushaplos(chromo, start, end)
}
\arguments{
\item{chromo}{Integer, chromosome number.}

\item{start}{Numeric, locus start position (Mb).}

\item{end}{Numeric, locus end position (Mb).}
}
\value{
region
}
\description{
Add haplotypes and generate allele scores at a locus of interest
}
\examples{
head(locushaplos(1,10000000,13000000))

}
